alerts and errors

+++ sys_manager.c
@@ -25,7 +25,7 @@ int available_workers = 0;
 typedef struct {
   int sens, cons, disp;
   Sensor *sensors;
-  int max_sensors, max_alerts;
+  Alert *alerts;
 } Mem_struct;
 
 int shmid;
@@ -93,7 +93,6 @@ void worker(int num) {
     available_workers++;
     sem_post(&BLOCK_WORKER);
   }
-
   exit(0);
 }
 
@@ -155,10 +154,9 @@ void cleanup() {
   for (int i = 0; i < N_WORKERS + 1; i++) {
     #ifdef DEBUG
       printf("DEBUG >> %d worker being deleted!\n", processes[i]);
-    #endif 
-    kill(processes[i], SIGKILL); // TODO: check if process is still doing something
+    #endif
+    if (processes[i] > 0) kill(processes[i], SIGKILL); // TODO: check if process is still doing something
   }
-
   for (int i = 0; i < N_WORKERS + 1; i++) wait(NULL);
 
   /* Remove Semaphores */
@@ -180,12 +178,16 @@ void cleanup() {
   pthread_cond_destroy(&cond);
 
   /* Remove shared memory */
-  shmctl(shmid, IPC_RMID, NULL); // remove the shared memory
-  shmdt(mem); // dettach from the shared memory
+  if (shmid > 0) {
+    shmctl(shmid, IPC_RMID, NULL); // remove the shared memory
+    shmdt(mem); // dettach from the shared memory
+  }
 
   /* Close log file */
-  write_log(fp, "HOME_IOT SIMULATOR CLOSING");
-  fclose(fp);
+  if (fp) {
+    write_log(fp, "HOME_IOT SIMULATOR CLOSING");
+    fclose(fp);
+  }
   exit(0);
 }
 
@@ -211,80 +213,87 @@ void sigtstp_handler(int sig) {
 int main(int argc, char **argv) {
 	if (argc != 2) {
     printf("INVALID NUMBER OF ARGUMENTS\n");
-    exit(0);
+    exit(EXIT_FAILURE);
   }
 
   fp = fopen(LOG_FILE, "a+");
   if (fp == NULL) {
     printf("ERROR >> OPENING LOG FILE\n");
-    exit(0);
+    exit(EXIT_FAILURE);
   }
 
   /* Handling file */
-	FILE *cfg = fopen(argv[1], "r");
-	if (cfg == NULL) {
+  FILE *cfg = fopen(argv[1], "r");
+  if (cfg == NULL) {
     printf("ERROR >> OPENING CONFIG FILE\n");
     fclose(cfg);
-		exit(0);
+		exit(EXIT_FAILURE);
 	}
 
 	/* Read Config */
-	fscanf(cfg, "%d\n%d\n%d\n%d\n%d", &QUEUE_SZ, &N_WORKERS, &MAX_KEYS, &MAX_SENSORS, &MAX_ALERTS);
-	if (QUEUE_SZ < 1 || N_WORKERS < 1 || MAX_KEYS < 1 || MAX_SENSORS < 1 || MAX_ALERTS < 0) {
+  fscanf(cfg, "%d\n%d\n%d\n%d\n%d", &QUEUE_SZ, &N_WORKERS, &MAX_KEYS, &MAX_SENSORS, &MAX_ALERTS);
+  if (QUEUE_SZ < 1 || N_WORKERS < 1 || MAX_KEYS < 1 || MAX_SENSORS < 1 || MAX_ALERTS < 0) {
     printf("ERROR >> IN CONFIG FILE\n");
     fclose(cfg);
-		exit(0);
-	}
+    exit(EXIT_FAILURE);
+  }
   fclose(cfg);
 
   /* Signal */
   act.sa_flags = 0;
-  sigemptyset(&act.sa_mask);
   // block all signals during setup
+  sigfillset(&act.sa_mask);
+  sigdelset(&act.sa_mask, SIGINT);
+  sigdelset(&act.sa_mask, SIGTSTP);
+  sigprocmask(SIG_SETMASK, &act.sa_mask, NULL); // this will block all signals
+
   act.sa_handler = SIG_IGN;
   sigaction(SIGINT, &act, NULL);
+  sigaction(SIGTSTP, &act, NULL);
 
   /* Mutexes */
   pthread_mutex_lock(&kill_mutex);
 
   if (pthread_cond_init(&cond, NULL) != 0) {
     printf("ERROR >> CREATING CONDITION VARIABLE\n");
-    exit(0);
+    exit(EXIT_FAILURE);
   }
 
   /* Semaphores */
   if (sem_init(&process_log, 0, 1) != 0) {
     printf("ERROR >> CREATING PROCESS_LOG SEMAPHORE\n");
-    exit(0);
+    exit(EXIT_FAILURE);
   }
 
   if (sem_init(&semaphore, 0, N_WORKERS) != 0) {
     printf("ERROR >> CREATING WORKERS SEMAPHORE\n");
-    exit(0);
+    exit(EXIT_FAILURE);
   }
 
   if (sem_init(&BLOCK_WORKER, 0, 1) != 0) {
     printf("ERROR >> CREATING BLOCK_WORKER SEMAPHORE\n");
-    exit(0);
+    exit(EXIT_FAILURE);
   }
 
   if ((sem_init(&BLOCK_SHM, 0, 1)) != 0) {
     printf("ERROR >> CREATING BLOCK_SHM SEMAPHORE\n");
-    exit(0);
+    exit(EXIT_FAILURE);
   }
 
   /* Main */
   write_log(fp, "HOME_IOT SIMULATOR STARTING");
 
   if ((processes = (pid_t *) malloc((N_WORKERS + 1) * sizeof(pid_t))) == NULL) {
-    printf("ERROR >> ALLOCATING MEMORY\n");
-    exit(0);
+    write_log(fp, "ERROR >> ALLOCATING MEMORY");
+    cleanup();
+    exit(EXIT_FAILURE);
   }
 
 	/* Shared Memory */
 	if ((shmid = shmget(IPC_PRIVATE, sizeof(Mem_struct*), 0666 | IPC_CREAT | IPC_EXCL)) == -1) {
-    printf("ERROR >> CREATING SHARED MEMORY\n");
-    exit(0);
+    write_log(fp, "ERROR >> CREATING SHARED MEMORY");
+    cleanup();
+    exit(EXIT_FAILURE);
 	}
 
   #ifdef DEBUG
@@ -292,36 +301,40 @@ int main(int argc, char **argv) {
   #endif
 
 	if ((mem = (Mem_struct *) shmat(shmid, NULL, 0)) == (Mem_struct *) -1) {
-    printf("ERROR >> ATTACHING SHARED MEMORY\n");
-		exit(0);
+    write_log(fp, "ERROR >> ATTACHING SHARED MEMORY");
+    cleanup();
+		exit(EXIT_FAILURE);
 	}
 
   sem_wait(&BLOCK_SHM);
   mem->cons = 0;
   mem->sens = 0;
   mem->disp = 0;
-  mem->max_sensors = MAX_SENSORS;
-  mem->max_alerts = MAX_ALERTS;
   mem->sensors = (Sensor *) malloc(MAX_SENSORS * sizeof(Sensor));
   for (int i = 0; i < MAX_SENSORS; i++) mem->sensors[i] = NULL_SENSOR;
+  mem->alerts = (Alert *) malloc(MAX_ALERTS * sizeof(Alert));
+  for (int i = 0; i < MAX_ALERTS; i++) mem->alerts[i] = NULL_ALERT;
   sem_post(&BLOCK_SHM);
 
   /* Sensor Reader */
   if (pthread_create(&sensor_reader, NULL, sensor_reader_func, NULL) != 0) {
-    printf("ERROR >> CREATING SENSOR_READER\n");
-    exit(0);
+    write_log(fp, "ERROR >> CREATING SENSOR_READER");
+    cleanup();
+    exit(EXIT_FAILURE);
   }
 
   /* Console Reader */
   if (pthread_create(&console_reader, NULL, console_reader_func, NULL) != 0) {
-    printf("ERROR >> CREATING CONSOLE_READER\n");
-    exit(0);
+    write_log(fp, "ERROR >> CREATING CONSOLE_READER");
+    cleanup();
+    exit(EXIT_FAILURE);
   }
 
   /* Dispatcher */
   if (pthread_create(&dispatcher, NULL, dispatcher_func, NULL) != 0) {
-    printf("ERROR >> CREATING DISPATCHER\n");
-    exit(0);
+    write_log(fp, "ERROR >> CREATING DISPATCHER");
+    cleanup();
+    exit(EXIT_FAILURE);
   }
 
   while (mem->sens == 0 || mem->cons == 0 || mem->disp == 0){}; // wait for threads to be ready
@@ -333,8 +346,9 @@ int main(int argc, char **argv) {
       else alert_watcher(); /* Alerts Watcher */
       exit(0);
     } else if (processes[i] < 0) {
-      printf("ERROR CREATING WORKER\n");
-      exit(0);
+      write_log(fp, "ERROR CREATING PROCESS");
+      cleanup();
+      exit(EXIT_FAILURE);
     }
   }
